requirements.txt
appdirs==1.4.3
distlib==0.3.0
dj-database-url==0.5.0
Django==2.2.3
django-classy-tags==1.0.0
django-cms==3.7.1
django-formtools==2.2
django-heroku==0.3.1
django-js-asset==1.2.2
django-mptt==0.6.0
django-reversion==3.0.7
django-sekizai==1.1.0
django-summernote==0.8.11.6
django-treebeard==4.3.1
djangocms-admin-style==1.5.0
djangocms-installer==1.2.1
djangocms-text-ckeditor==3.9.0
filelock==3.0.12
gunicorn==20.0.4
html5lib==1.0.1
importlib-metadata==1.5.0
mysqlclient==1.4.6
Pillow==7.0.0
psycopg2==2.8.4
pycurl==7.43.0.5
PyMySQL==0.9.3
pytz==2019.3
six==1.14.0
South==1.0.2
sqlparse==0.3.1
tzlocal==2.0.0
virtualenv==20.0.13
webencodings==0.5.1
whitenoise==5.0.1
zipp==3.1.0
